backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.1.0
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
httpx==0.25.2
beautifulsoup4==4.12.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
asyncio==3.4.3
logging==0.4.9.6 